requirements.txt
Flask
Flask-SocketIO
Flask-SQLAlchemy
python-socketio[client]
InquirerPy
requests
pyyaml
typer